forex_trading_module/requirements.txt
yfinance>=0.2.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.6.0
sqlalchemy>=2.0.0
scipy>=1.10.0
python-dotenv>=1.0.0
feedparser>=6.0.0
requests>=2.28.0